+++ functions/sitemap.xml.ts
@@ -41,6 +41,17 @@ export async function onRequest(context: any) {
     "sman-1-jakarta", "sman-8-jakarta", "smpn-1-surabaya", "sdn-1-bandung", "sman-3-malang", "smkn-1-denpasar", "man-2-yogyakarta"
   ];
 
+  const provinces = [
+    "aceh", "sumatera-utara", "sumatera-barat", "riau", "jambi", "sumatera-selatan", "bengkulu", "lampung", "kepulauan-bangka-belitung", "kepulauan-riau",
+    "dki-jakarta", "jawa-barat", "jawa-tengah", "di-yogyakarta", "jawa-timur", "banten",
+    "bali", "nusa-tenggara-barat", "nusa-tenggara-timur",
+    "kalimantan-barat", "kalimantan-tengah", "kalimantan-selatan", "kalimantan-timur", "kalimantan-utara",
+    "sulawesi-utara", "sulawesi-tengah", "sulawesi-selatan", "sulawesi-tenggara", "gorontalo", "sulawesi-barat",
+    "maluku", "maluku-utara",
+    "papua", "papua-barat", "papua-selatan", "papua-tengah", "papua-pegunungan", "papua-barat-daya"
+  ];
+
+
   let xml = `<?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
@@ -53,6 +64,15 @@ export async function onRequest(context: any) {
     xml += `
   <url>
     <loc>${baseUrl}/kota/${city}</loc>
+    <changefreq>weekly</changefreq>
+    <priority>0.8</priority>
+  </url>`;
+  }
+
+  for (const prov of provinces) {
+    xml += `
+  <url>
+    <loc>${baseUrl}/provinsi/${prov}</loc>
     <changefreq>weekly</changefreq>
     <priority>0.8</priority>
   </url>`;